src/test/kotlin/com/clouddjr/advent2023/Day21Test.kt
package com.clouddjr.advent2023

import com.clouddjr.advent2023.Resources.resourceAsListOfString
import org.junit.jupiter.api.DisplayName
import org.junit.jupiter.api.Nested
import org.junit.jupiter.api.Test
import kotlin.test.assertEquals

@DisplayName("Day 21")
class Day21Test {
    @Nested
    @DisplayName("Part 1")
    inner class Part1 {
        @Test
        fun `Matches actual`() {
            val answer = Day21(resourceAsListOfString("day21.txt")).solvePart1()

            assertEquals(3594, answer)
        }
    }

    @Nested
    @DisplayName("Part 2")
    inner class Part2 {
        @Test
        fun `Matches actual`() {
            val answer = Day21(resourceAsListOfString("day21.txt")).solvePart2()

            assertEquals(605_247_138_198_755, answer)
        }
    }
}